requirements.txt
spotipy==2.23.0
qrcode==7.4.2
Pillow==10.1.0
reportlab==4.0.7
python-dotenv==1.0.0

